feat: enhance LLM configuration and implement CrawlerService for improved feed processing
