+++ src/Crawler/appsettings.json
@@ -13,13 +13,22 @@
     "maxAgeDays":  7
   },
   "llm": {
-    "provider": "openai",
-    "model": "openai/gpt-5-mini",
+    "provider": "github",
     "maxCallsPerRun": 80,
-  "minCharsToSummarize": 500,
-  "requestsPerWindow": 3,
-  "windowSeconds": 60,
-  "initialDelaySeconds": 2
+    "github": {
+      "minCharsToSummarize": 500,
+      "requestsPerWindow": 3,
+      "windowSeconds": 60,
+      "initialDelaySeconds": 2,
+      "model": "openai/gpt-5-mini"
+    },
+    "openai": {
+      "minCharsToSummarize": 500,
+      "requestsPerWindow": 3,
+      "windowSeconds": 60,
+      "initialDelaySeconds": 2,
+      "model": "openai/gpt-5-mini"
+    }
   },
   "feeds": {
     "dataFile": "feed-config/feeds.json"
